+++ importer/specobject/src/main/java/module-info.java
@@ -1,15 +1,14 @@
 /**
  * This provides an importer for the SpecObject XML format.
+ * 
+ * @provides org.itsallcode.openfasttrace.api.importer.ImporterFactory
  */
 module org.itsallcode.openfasttrace.importer.specobject
 {
-    exports org.itsallcode.openfasttrace.importer.specobject;
-    exports org.itsallcode.openfasttrace.importer.specobject.handler;
-    exports org.itsallcode.openfasttrace.importer.specobject.xml;
-    exports org.itsallcode.openfasttrace.importer.specobject.xml.tree;
-    exports org.itsallcode.openfasttrace.importer.specobject.xml.event;
-
     requires java.logging;
     requires transitive java.xml;
     requires transitive org.itsallcode.openfasttrace.api;
+
+    provides org.itsallcode.openfasttrace.api.importer.ImporterFactory
+            with org.itsallcode.openfasttrace.importer.specobject.SpecobjectImporterFactory;
 }