+++ importer/tag/src/main/java/module-info.java
@@ -1,10 +1,13 @@
 /**
  * This provides an importer for coverage tags.
+ * 
+ * @provides org.itsallcode.openfasttrace.api.importer.ImporterFactory
  */
 module org.itsallcode.openfasttrace.importer.tag
 {
-    exports org.itsallcode.openfasttrace.importer.tag;
-
     requires java.logging;
     requires transitive org.itsallcode.openfasttrace.api;
+
+    provides org.itsallcode.openfasttrace.api.importer.ImporterFactory
+            with org.itsallcode.openfasttrace.importer.tag.TagImporterFactory;
 }