+++ importer/zip/src/main/java/module-info.java
@@ -1,10 +1,12 @@
 /**
  * This provides an importer for the ZIP files.
+ * 
+ * @provides org.itsallcode.openfasttrace.api.importer.ImporterFactory
  */
 module org.itsallcode.openfasttrace.importer.zip
 {
-    exports org.itsallcode.openfasttrace.importer.zip;
-    exports org.itsallcode.openfasttrace.importer.zip.input;
-
     requires transitive org.itsallcode.openfasttrace.api;
+
+    provides org.itsallcode.openfasttrace.api.importer.ImporterFactory
+            with org.itsallcode.openfasttrace.importer.zip.ZipFileImporterFactory;
 }