+++ reporter/html/src/main/java/module-info.java
@@ -1,11 +1,12 @@
 /**
  * This provides an report generator for the HTML format.
+ * 
+ * @provides org.itsallcode.openfasttrace.api.report.ReporterFactory
  */
 module org.itsallcode.openfasttrace.report.html
 {
-    exports org.itsallcode.openfasttrace.report.html;
-    exports org.itsallcode.openfasttrace.report.html.view;
-    exports org.itsallcode.openfasttrace.report.html.view.html;
-
     requires transitive org.itsallcode.openfasttrace.api;
+
+    provides org.itsallcode.openfasttrace.api.report.ReporterFactory
+            with org.itsallcode.openfasttrace.report.html.HtmlReporterFactory;
 }